+++ core/tools.js
@@ -1807,6 +1807,45 @@
 				for ( i = 0; i < len; i++ ) {
 					fn.call( thisArg, array[ i ], i, array );
 				}
+			},
+
+			/**
+			 * Applies a function to each element of an array and returns the array of results in the same order.
+			 * Note the order of the parameters.
+			 *
+			 * @param {Array} array Array over elements of which `fn` is applied.
+			 * @param {Function} fn A function with the signature `a -> b`.
+			 * @param {Mixed} [thisArg=undefined] The context object for `fn`.
+			 * @returns {Array} Array of mapped elements.
+			 * @member CKEDITOR.tools.array
+			 * @since 4.6.2
+			 */
+			map: function( array, fn, thisArg ) {
+				var result = [];
+				for ( var i = 0; i < array.length; i++ ) {
+					result.push( fn.call( thisArg, array[ i ], i, array ) );
+				}
+				return result;
+			},
+
+			/**
+			 * Applies a function against each value in an array storing the result in an accumulator passed to the next iteration.
+			 * Note the order of the parameters.
+			 *
+			 * @param {Array} array Array over elements of which `fn` is applied.
+			 * @param {Function} fn A function with the signature `(accumulator, a, index, array) -> b`.
+			 * @param {Mixed} initial Initial value of the accumulator.
+			 * @param {Mixed} [thisArg=undefined] The context object for `fn`.
+			 * @returns {Mixed} Final value of the accumulator.
+			 * @member CKEDITOR.tools.array
+			 * @since 4.6.2
+			*/
+			reduce: function( array, fn, initial, thisArg ) {
+				var acc = initial;
+				for ( var i = 0; i < array.length; i++ ) {
+					acc = fn.call( thisArg, acc, array[ i ], i, array );
+				}
+				return acc;
 			}
 		}
 	};